rename: @/assets/icon 디렉토리 생성
